Change enum Status to Scene

+++ headers/GamePlay/Level.h
@@ -1,7 +1,7 @@
 #include "../Moving.h"
 
 struct LevelScene {
-    unsigned int count;
+    unsigned short int count;
     Color background;
     TextBox TextBoxs[5];
     Selector1D selector = {1};
